feat: Added user model

+++ src/sockets/room.events.ts
@@ -1,7 +1,9 @@
 import app from '../app';
-import {namespaces} from './namespaces';
+import {isNil} from 'ramda';
+import {namespaces, events} from './namespaces';
 import logger from '../util/logger';
 import {Room} from '../models/Room.model';
+import {User} from '../models/User.model';
 
 export const roomSocketEvents = () => {
     const io = app.get('io');
@@ -12,19 +14,25 @@ export const roomSocketEvents = () => {
     roomSocket.on('connection', (socket: any) => {
         console.info(`socket-io room: connected`);
 
-        socket.on('SOCKET/REQUEST_ROOM_ID', async () => {
+        socket.on(events.REQUEST_ROOM_ID, async () => {
             roomId = Math.round(Math.random() * 10);
             await new Room({roomId}).save();
 
-            logger.info(`io: REQUEST_ROOM_ID -> id : ${roomId}`);
+            logger.info(`io: ${events.REQUEST_ROOM_ID} -> id : ${roomId}`);
         
-            socket.broadcast.emit('SOCKET/GET_ROOM_ID', roomId);
+            socket.broadcast.emit(events.GET_ROOM_ID, roomId);
+        });
+
+        socket.on(events.IS_PSEUDO_AVAILABLE, async ({pseudo}: {pseudo: string}) => {
+            const user = await User.findOne({pseudo});
+            isNil(user) && socket.broadcast.emit(events.IS_PSEUDO_AVAILABLE_STATUS, true);
         });
 
         socket.on('disconnect', async () => {
-            logger.info(`room: disconnected: REQUEST_ROOM_ID -> id : ${roomId}`);
-            await Room.deleteOne({roomId});
-            console.log('socket-io ');
+            if (roomId) {
+                logger.info(`room: disconnected: ${events.REQUEST_ROOM_ID} -> id : ${roomId}`);
+                await Room.deleteOne({roomId});
+            }
         });
     });
 }
